Create standalone serializer and deserializer
Removed the Message wrapper from the earlier implementation and created
the serializer and deserializer as standalone classes. The interface and
type classes need to be annotated with @JsonSerialize and
@JsonDeserialize. It's not a complete removal of Jackson from the type
classes but it's far tidier than sprinkling them throughout the
generated code and removes the need for the static serializer and
deserializer classes that appeared in every class representing a complex
type.