Add de-serialization support (#1)
* Add de-serialization support

Start with a tidy up:

- add some basic Javadoc
- rename classes to denote their JSON-based functionality
- move the car types into their own sub-package

* Add @JsonDeserialize annotation

* More fun with deserializers and type references

* Type creation in the deserializer

Added the ContextualDeserializer interface to cache the JavaType for use
in the deserialize method.

* Initial deserialization call plus refactoring

The message derserializer creates the top level payload object and
passes it off to the helper method that walks the JSON in conjunction
with the property lists from the serializable payload type. Also
refactored the package paths to make the de-coupling clearer regarding
the visibility of classes in the compilation and runtime modules.

* Quick fix in ListProperty deserialization processing

Code was checking root.isArray() instead of field.isArray() where field
is the entry found when searching root.

* Modify package paths

Back towards what I started with after chatting with Alastair and
realising that the code isn't split exactly as I thought it was.

* Remove serialization classes from Message

Split out the MessageJsonSerializer and MessageJsonDeserializer into two
standalone classes that include their static helper methods.

* Quick clean up of the surrogates for generated types

* Create standalone serializer and deserializer

Removed the Message wrapper from the earlier implementation and created
the serializer and deserializer as standalone classes. The interface and
type classes need to be annotated with @JsonSerialize and
@JsonDeserialize. It's not a complete removal of Jackson from the type
classes but it's far tidier than sprinkling them throughout the
generated code and removes the need for the static serializer and
deserializer classes that appeared in every class representing a complex
type.

* Delete unused CustomTypeResolverBuilder class

* Rearranged test code

Moved the surrogates for our generated types into the src/test/java
source folder and added testPlayground as a replacement for testing in
the static main method. Also added serialization support for boolean
along with a property in SimpleCar to allow it to be tested.

* Functional serialization and deserialization

First version that is functional in both directions. Works with complex
objects (SerializablePayload), simple values (String, Integer, Double,
Boolean), collections of complex objects or simple values, and supports
nulls.

Could do with a few more tests but those that are present are pretty
useful.