Switch to an annotation processing
The abstract methods from the SerializablePayload interface and the
objects that supported them were becoming cumbersome to the point that
it was making it very difficult to implement some aspects of the generic
serialization components.

The decision was was to switch to adding custom annotations to the type
object classes. The annotation data is then used at runtime to decide
which types and fields are to be processed and how they're to be
encoded. The new additions are listed below.

Annotations:

- @TypeDefinition
   - @FieldOrder
   - @CustomTypeEncoding
- @TypeProperty

Enums:

- CustomTypeEncoder
- FieldTypeEncoding