Performance improvements
This change adds two caches to the SerializationHelper (formerly known
as SerializationUtil):

- a map of Class<?> to field name, Field, and AwsFieldProperty
- a map of Class<?> to boolean result of usesCustomTypeEncoding

The regular deserialization has been changed to iterate over the nodes
that are present rather than the Field objects that might be present
(then searching for the node).

There was also a bug in the custom serializer in the case where it was
called with a String. The instanceof check was being executed after the
serialize method had already determined that it was dealing with a
complex type.